backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
web3
pydantic
httpx
pytest
opencv-python
ultralytics
numpy
scipy
